// resources/js/types/data.d.ts
export type DataRecord = {
    id: number;
};

type PaginationLink = {
    url: string;
    label: string;
    active: boolean;
};

export type Pagination<T> = {
    current_page: number;
    data: T[];
    first_page_url: string;
    from: number;
    last_page: number;
    last_page_url: string;
    links: PaginationLink[];
    next_page_url: string;
    path: string;
    per_page: number;
    prev_page_url: string;
    to: number;
    total: number;
};